+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - #22 Store validation errors to the form
 - #13 Add CallbackFilter
+### Deprecated
+- #26 Deprecate `\Fiv\Form\Filter\Trim`
 
 ## 0.1.0-alpha3 [2016-03-25]